Update GUI
Refactor Amiya response and exit flow for GUI integration

The system was previously unable to show Amiya's responses for more complex inputs
on the GUI and the exit message was not
displayed before the application closed.

Let's
* change the return type of Ui methods to strings,
so that getResponse can directly use these
* modify the init of Amiya
* use asynchronous programming in Java to
guarantee the message is shown before exiting.

This approach enhances the user experience by ensuring a smooth transition
and proper display of the exit message before application termination.